+++ src/ONTraC/run/processes.py
@@ -3,14 +3,13 @@ from typing import Callable, Dict
 
 import numpy as np
 
-from ..data import load_dataset
 from ..GNN import evaluate, predict, save_graph_pooling_results, set_seed, train
 from ..log import *
 from ..model import GNN
 from ..niche_net import construct_niche_network, gen_samples_yaml
-from ..niche_trajectory import NTScore_table, get_niche_NTScore, load_consolidate_data, niche_to_cell_NTScore
+from ..niche_trajectory import NTScore_table, get_niche_NTScore, niche_to_cell_NTScore
 from ..optparser import *
-from ..preprocessing.pp_control import preprocessing_gnn, preprocessing_nn
+from ..preprocessing.pp_control import preprocessing_gnn, preprocessing_nn, preprocessing_nt
 from ..train import GNNBatchTrain
 from ..train.inspect_funcs import get_inspect_funcs
 from ..utils import get_rel_params, read_yaml_file
@@ -93,7 +92,9 @@ def gnn(options: Values) -> None:
     # evaluate
     evaluate(batch_train=batch_train)
     # predict
-    consolidate_z_array, consolidate_s_array, _ = predict(output_dir=options.GNN_dir, batch_train=batch_train, dataset=dataset)
+    consolidate_z_array, consolidate_s_array, _ = predict(output_dir=options.GNN_dir,
+                                                          batch_train=batch_train,
+                                                          dataset=dataset)
     # save results
     if consolidate_z_array is not None and consolidate_s_array is not None:
         save_graph_pooling_results(meta_data_df=meta_data_df,
@@ -114,25 +115,26 @@ def niche_trajectory_construct(options: Values) -> None:
     """
 
     info('----------------- Niche trajectory ------------------ ')
-    consolidate_s_array, consolidate_out_adj_array = load_consolidate_data(GNN_dir=options.GNN_dir)
+    # load data
+    meta_data_df, niche_level_niche_cluster_assign_df, consolidate_out_adj_array = preprocessing_nt(
+        NN_dir=options.NN_dir, GNN_dir=options.GNN_dir)
 
     params = read_yaml_file(f'{options.NN_dir}/samples.yaml')
     rel_params = get_rel_params(NN_dir=options.NN_dir, params=params)
-    dataset = load_dataset(NN_dir=options.NN_dir)
 
-    niche_cluster_score, niche_level_NTScore = get_niche_NTScore(
+    niche_cluster_score, niche_level_NTScore_df = get_niche_NTScore(
         trajectory_construct_method=options.trajectory_construct,
-        niche_cluster_loading=consolidate_s_array,
+        niche_level_niche_cluster_assign_df=niche_level_niche_cluster_assign_df,
         niche_adj_matrix=consolidate_out_adj_array)
-    cell_level_NTScore, all_niche_level_NTScore_dict, all_cell_level_NTScore_dict = niche_to_cell_NTScore(
-        dataset=dataset, rel_params=rel_params, niche_level_NTScore=niche_level_NTScore)
-
     np.savetxt(fname=f'{options.NT_dir}/niche_cluster_score.csv.gz', X=niche_cluster_score, delimiter=',')
-    np.savetxt(fname=f'{options.NT_dir}/niche_NTScore.csv.gz', X=niche_level_NTScore, delimiter=',')
-    np.savetxt(fname=f'{options.NT_dir}/cell_NTScore.csv.gz', X=cell_level_NTScore, delimiter=',')
+
+    cell_level_NTScore_df = niche_to_cell_NTScore(meta_data_df=meta_data_df,
+                                                  niche_level_NTScore_df=niche_level_NTScore_df,
+                                                  rel_params=rel_params)
 
     NTScore_table(save_dir=options.NT_dir,
-                  rel_params=rel_params,
-                  all_niche_level_NTScore_dict=all_niche_level_NTScore_dict,
-                  all_cell_level_NTScore_dict=all_cell_level_NTScore_dict)
+                  meta_data_df=meta_data_df,
+                  niche_level_NTScore_df=niche_level_NTScore_df,
+                  cell_level_NTScore_df=cell_level_NTScore_df,
+                  rel_params=rel_params)
     info('--------------- Niche trajectory end ---------------- ')